Nova adaptação feita
